Alexa sync state report (#24835)
* Do a sync after changing state reporting

* Fix entity config being None
