+++ tests/components/cloud/conftest.py
@@ -1,9 +1,10 @@
 """Fixtures for cloud tests."""
-import pytest
-
 from unittest.mock import patch
 
-from homeassistant.components.cloud import prefs
+import jwt
+import pytest
+
+from homeassistant.components.cloud import const, prefs
 
 from . import mock_cloud, mock_cloud_prefs
 
@@ -28,3 +29,19 @@ async def cloud_prefs(hass):
     cloud_prefs = prefs.CloudPreferences(hass)
     await cloud_prefs.async_initialize()
     return cloud_prefs
+
+
+@pytest.fixture
+async def mock_cloud_setup(hass):
+    """Set up the cloud."""
+    await mock_cloud(hass)
+
+
+@pytest.fixture
+def mock_cloud_login(hass, mock_cloud_setup):
+    """Mock cloud is logged in."""
+    hass.data[const.DOMAIN].id_token = jwt.encode({
+        'email': 'hello@home-assistant.io',
+        'custom:sub-exp': '2018-01-03',
+        'cognito:username': 'abcdefghjkl',
+    }, 'test')
